documentation; moved files
