+++ src/demo/index.css
@@ -1,4 +1,4 @@
-@import "../node_modules/milligram/dist/milligram.css";
+@import "../../node_modules/milligram/dist/milligram.css";
 @import "../theme.css";
 
 html {
