+++ src/index.scss
@@ -1,5 +1,5 @@
 @import "../node_modules/milligram/dist/milligram.css";
-@import "../theme.css";
+@import "theme.css";
 
 body {
   background: var(--color-primary-1);
