+++ src/theme.css
@@ -1,0 +1,29 @@
+:root {
+  --color-primary-0: #E7E392; /* Main Primary color */
+  --color-primary-1: #FFFDD7;
+  --color-primary-2: #FFFCBE;
+  --color-primary-3: #CEC968;
+  --color-primary-4: #B2AC3F;
+
+  --color-secondary-1-0: #67A483; /* Main Secondary color (1) */
+  --color-secondary-1-1: #AACAB9;
+  --color-secondary-1-2: #86B59C;
+  --color-secondary-1-3: #49926B;
+  --color-secondary-1-4: #2C7E52;
+
+  --color-secondary-2-0: #E7C192; /* Main Secondary color (2) */
+  --color-secondary-2-1: #FFEDD7;
+  --color-secondary-2-2: #FFE2BE;
+  --color-secondary-2-3: #CEA168;
+  --color-secondary-2-4: #B27F3F;
+
+  --color-complement-0: #68779B; /* Main Complement color */
+  --color-complement-1: #A8B0C3;
+  --color-complement-2: #8490AC;
+  --color-complement-3: #4D5F8B;
+  --color-complement-4: #334678;
+
+  --white: #FFFFFF;
+  --black: #000000;
+  --dark: #202041;
+}
